Fix operation extractor for Collection parameters
