+++ src/.msbuild/version.props
@@ -4,7 +4,7 @@
         <!-- major.minor -->
         <TwoDigitVersion Condition="'$(TwoDigitVersion)'==''">2.0</TwoDigitVersion>
         
-        <PatchNum Condition="'$(PatchNum)'==''">2</PatchNum>
+        <PatchNum Condition="'$(PatchNum)'==''">3</PatchNum>
         <!-- special postfix will be added to patch num if defined -->
         <PatchNumSuffix Condition="'$(PatchNumSuffix)'=='' And '$(Configuration)'!='Release'">debug</PatchNumSuffix>
         <BuildNum Condition="'$(BuildNum)'==''">0</BuildNum>
